requirements.txt
tqdm>=4.62.3
pyyaml>=5.4.1
tensorboard>=2.7.0
torch>=1.6.0
